refactor_javabean_postion
