+++ dinky-admin/src/main/java/org/dinky/controller/AlertInstanceController.java
@@ -21,7 +21,7 @@ package org.dinky.controller;
 
 import org.dinky.alert.AlertPool;
 import org.dinky.alert.AlertResult;
-import org.dinky.data.annotation.Log;
+import org.dinky.data.annotations.Log;
 import org.dinky.data.constant.PermissionConstants;
 import org.dinky.data.dto.AlertInstanceDTO;
 import org.dinky.data.enums.BusinessType;
