+++ dinky-admin/src/main/java/org/dinky/controller/JobInstanceController.java
@@ -21,7 +21,7 @@ package org.dinky.controller;
 
 import org.dinky.api.FlinkAPI;
 import org.dinky.assertion.Asserts;
-import org.dinky.data.annotation.Log;
+import org.dinky.data.annotations.Log;
 import org.dinky.data.enums.BusinessType;
 import org.dinky.data.model.ID;
 import org.dinky.data.model.JobInfoDetail;
