+++ dinky-admin/src/main/java/org/dinky/controller/MenuController.java
@@ -19,7 +19,7 @@
 
 package org.dinky.controller;
 
-import org.dinky.data.annotation.Log;
+import org.dinky.data.annotations.Log;
 import org.dinky.data.constant.PermissionConstants;
 import org.dinky.data.dto.MenuDTO;
 import org.dinky.data.dto.RoleMenuDTO;
