+++ dinky-admin/src/main/java/org/dinky/controller/TaskVersionController.java
@@ -19,7 +19,7 @@
 
 package org.dinky.controller;
 
-import org.dinky.data.annotation.Log;
+import org.dinky.data.annotations.Log;
 import org.dinky.data.dto.TaskVersionHistoryDTO;
 import org.dinky.data.enums.BusinessType;
 import org.dinky.data.model.TaskVersion;
