+++ dinky-admin/src/main/java/org/dinky/controller/UserController.java
@@ -20,7 +20,7 @@
 package org.dinky.controller;
 
 import org.dinky.assertion.Asserts;
-import org.dinky.data.annotation.Log;
+import org.dinky.data.annotations.Log;
 import org.dinky.data.constant.PermissionConstants;
 import org.dinky.data.dto.ModifyPasswordDTO;
 import org.dinky.data.enums.BusinessType;
