--- dinky-admin/src/main/java/org/dinky/data/annotation/ConditionalOnListProperty.java
@@ -1,58 +1,0 @@
-/*
- *
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package org.dinky.data.annotation;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import java.util.Arrays;
-
-import org.springframework.context.annotation.Condition;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.context.annotation.Conditional;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Conditional(CustomListCondition.class)
-public @interface ConditionalOnListProperty {
-    String name();
-
-    String havingValue();
-}
-
-class CustomListCondition implements Condition {
-
-    @Override
-    public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
-        String propertyName = (String) metadata.getAnnotationAttributes(ConditionalOnListProperty.class.getName())
-                .get("name");
-        String requiredValue = (String) metadata.getAnnotationAttributes(ConditionalOnListProperty.class.getName())
-                .get("havingValue");
-        String[] propertyValues = context.getEnvironment().getProperty(propertyName, String[].class);
-        if (propertyValues != null) {
-            return Arrays.asList(propertyValues).contains(requiredValue);
-        }
-        return false;
-    }
-}
