+++ dinky-common/src/main/java/org/dinky/data/annotations/SupportDialect.java
@@ -17,9 +17,9 @@
  *
  */
 
-package org.dinky.data.annotation;
+package org.dinky.data.annotations;
 
-import org.dinky.data.enums.BusinessType;
+import org.dinky.config.Dialect;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -27,17 +27,15 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** 自定义操作日志记录注解 */
-@Target({ElementType.PARAMETER, ElementType.METHOD})
+/**
+ * task support dialect
+ */
+@Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface Log {
-    /** 模块 */
-    public String title() default "";
-
-    /** 功能 */
-    public BusinessType businessType() default BusinessType.OTHER;
-
-    /** 是否保存请求的参数 */
-    public boolean isSaveRequestData() default true;
+public @interface SupportDialect {
+    /**
+     * Support Dialect
+     */
+     Dialect[] value();
 }
