+++ dinky-common/src/main/java/org/dinky/data/constant/BaseConstant.java
