+++ dinky-common/src/main/java/org/dinky/data/constant/PermissionConstants.java
