+++ dinky-common/src/main/java/org/dinky/data/enums/BusinessType.java
