+++ dinky-common/src/main/java/org/dinky/data/enums/CodeEnum.java
